# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Sabaik Almasa API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResult {
  success: boolean;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface AdminUser {
  id: number;
  username: string;
  name: string;
}

export interface AuthResult {
  token: string;
  user: AdminUser;
}

export interface HeroSlide {
  id: number;
  title: string;
  subtitle: string;
  imageUrl: string;
  /** @nullable */
  ctaText?: string | null;
  order: number;
  isActive: boolean;
  createdAt?: string;
}

export interface HeroSlideInput {
  title: string;
  subtitle: string;
  imageUrl: string;
  /** @nullable */
  ctaText?: string | null;
  order: number;
  isActive?: boolean;
}

export interface Service {
  id: number;
  title: string;
  description: string;
  icon: string;
  /** @nullable */
  imageUrl?: string | null;
  order: number;
  isActive: boolean;
}

export interface ServiceInput {
  title: string;
  description: string;
  icon: string;
  /** @nullable */
  imageUrl?: string | null;
  order: number;
  isActive?: boolean;
}

export interface Container {
  id: number;
  name: string;
  size: string;
  capacity: string;
  description: string;
  features: string[];
  pricePerDay: number;
  imageUrl: string;
  order: number;
  isActive: boolean;
}

export interface ContainerInput {
  name: string;
  size: string;
  capacity: string;
  description: string;
  features: string[];
  pricePerDay: number;
  imageUrl: string;
  order: number;
  isActive?: boolean;
}

export interface CompanyValue {
  id: number;
  title: string;
  description: string;
  icon: string;
  order: number;
}

export interface CompanyValueInput {
  title: string;
  description: string;
  icon: string;
  order: number;
}

export interface Testimonial {
  id: number;
  clientName: string;
  company: string;
  content: string;
  rating: number;
  /** @nullable */
  avatarUrl?: string | null;
  isActive: boolean;
  createdAt?: string;
}

export interface TestimonialInput {
  clientName: string;
  company: string;
  content: string;
  rating: number;
  /** @nullable */
  avatarUrl?: string | null;
  isActive?: boolean;
}

export interface Partner {
  id: number;
  name: string;
  logoUrl: string;
  /** @nullable */
  websiteUrl?: string | null;
  order: number;
  isActive: boolean;
}

export interface PartnerInput {
  name: string;
  logoUrl: string;
  /** @nullable */
  websiteUrl?: string | null;
  order: number;
  isActive?: boolean;
}

export type ServiceRequestStatus = typeof ServiceRequestStatus[keyof typeof ServiceRequestStatus];


export const ServiceRequestStatus = {
  pending: 'pending',
  in_progress: 'in_progress',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface ServiceRequest {
  id: number;
  clientName: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  serviceType: string;
  containerSize: string;
  location: string;
  /** @nullable */
  duration?: string | null;
  /** @nullable */
  notes?: string | null;
  status: ServiceRequestStatus;
  /** @nullable */
  adminNotes?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export interface ServiceRequestInput {
  clientName: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  serviceType: string;
  containerSize: string;
  location: string;
  /** @nullable */
  duration?: string | null;
  /** @nullable */
  notes?: string | null;
}

export type ServiceRequestUpdateStatus = typeof ServiceRequestUpdateStatus[keyof typeof ServiceRequestUpdateStatus];


export const ServiceRequestUpdateStatus = {
  pending: 'pending',
  in_progress: 'in_progress',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface ServiceRequestUpdate {
  status?: ServiceRequestUpdateStatus;
  /** @nullable */
  adminNotes?: string | null;
}

export interface DashboardStats {
  totalRequests: number;
  pendingRequests: number;
  inProgressRequests: number;
  completedRequests: number;
  totalConversations: number;
  openConversations: number;
  unreadNotifications: number;
  recentRequests: ServiceRequest[];
}

export type ConversationStatus = typeof ConversationStatus[keyof typeof ConversationStatus];


export const ConversationStatus = {
  open: 'open',
  in_progress: 'in_progress',
  closed: 'closed',
} as const;

export interface Conversation {
  id: number;
  clientName: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  subject?: string | null;
  status: ConversationStatus;
  /** @nullable */
  lastMessage?: string | null;
  unreadCount?: number;
  createdAt: string;
  updatedAt: string;
}

export interface ConversationInput {
  clientName: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  subject?: string | null;
}

export type ConversationUpdateStatus = typeof ConversationUpdateStatus[keyof typeof ConversationUpdateStatus];


export const ConversationUpdateStatus = {
  open: 'open',
  in_progress: 'in_progress',
  closed: 'closed',
} as const;

export interface ConversationUpdate {
  status?: ConversationUpdateStatus;
}

export type MessageSenderType = typeof MessageSenderType[keyof typeof MessageSenderType];


export const MessageSenderType = {
  client: 'client',
  admin: 'admin',
  ai: 'ai',
} as const;

export interface Message {
  id: number;
  conversationId: number;
  content: string;
  senderType: MessageSenderType;
  isRead: boolean;
  createdAt: string;
}

export type MessageInputSenderType = typeof MessageInputSenderType[keyof typeof MessageInputSenderType];


export const MessageInputSenderType = {
  client: 'client',
  admin: 'admin',
  ai: 'ai',
} as const;

export interface MessageInput {
  content: string;
  senderType: MessageInputSenderType;
}

export interface AiMessageInput {
  message: string;
  /** @nullable */
  conversationId?: number | null;
}

export interface BotReply {
  reply: string;
  /** @nullable */
  conversationId?: number | null;
}

export type NotificationType = typeof NotificationType[keyof typeof NotificationType];


export const NotificationType = {
  service_request: 'service_request',
  chat: 'chat',
  system: 'system',
} as const;

export interface Notification {
  id: number;
  title: string;
  message: string;
  type: NotificationType;
  isRead: boolean;
  /** @nullable */
  refId?: number | null;
  /** @nullable */
  refType?: string | null;
  createdAt: string;
}

export type GetServiceRequestsParams = {
status?: GetServiceRequestsStatus;
};

export type GetServiceRequestsStatus = typeof GetServiceRequestsStatus[keyof typeof GetServiceRequestsStatus];


export const GetServiceRequestsStatus = {
  pending: 'pending',
  in_progress: 'in_progress',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;