# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sabaik Almasa API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
...
  /services:
    get:
      operationId: getServices
      tags: [content]
      summary: List services
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [content]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{id}:
    patch:
      operationId: updateService
      tags: [content]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [content]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── Containers ────────────────────────────────────────────────────────────
  /containers:
    get:
      operationId: getContainers
      tags: [content]
      summary: List container types
      responses:
        "200":
          description: List of containers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Container"
    post:
      operationId: createContainer
      tags: [content]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContainerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Container"
  /containers/{id}:
    patch:
      operationId: updateContainer
      tags: [content]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContainerInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Container"
    delete:
      operationId: deleteContainer
      tags: [content]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── Values ────────────────────────────────────────────────────────────────
...
  /testimonials:
    get:
      operationId: getTestimonials
      tags: [content]
      summary: List testimonials
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Testimonial"
    post:
      operationId: createTestimonial
      tags: [content]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestimonialInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Testimonial"
  /testimonials/{id}:
    patch:
      operationId: updateTestimonial
      tags: [content]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestimonialInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Testimonial"
    delete:
      operationId: deleteTestimonial
      tags: [content]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── Partners ──────────────────────────────────────────────────────────────
...
  /service-requests:
    get:
      operationId: getServiceRequests
      tags: [services]
      summary: List all service requests (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, in_progress, completed, cancelled]
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceRequest"
    post:
      operationId: submitServiceRequest
      tags: [services]
      summary: Submit a new service request (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceRequestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequest"
  /service-requests/{id}:
    get:
      operationId: getServiceRequest
      tags: [services]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Service request detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequest"
    patch:
      operationId: updateServiceRequest
      tags: [services]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceRequestUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequest"

  # ── Dashboard Stats ───────────────────────────────────────────────────────
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }
    AuthResult:
      type: object
      required: [token, user]
      properties:
        token: { type: string }
        user:
          $ref: "#/components/schemas/AdminUser"
    AdminUser:
      type: object
      required: [id, username, name]
      properties:
        id: { type: integer }
        username: { type: string }
        name: { type: string }
...
    Service:
      type: object
      required: [id, title, description, icon, order, isActive]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string }
        icon: { type: string }
        imageUrl: { type: ["string", "null"] }
        order: { type: integer }
        isActive: { type: boolean }
    ServiceInput:
      type: object
      required: [title, description, icon, order]
      properties:
        title: { type: string }
        description: { type: string }
        icon: { type: string }
        imageUrl: { type: ["string", "null"] }
        order: { type: integer }
        isActive: { type: boolean }
    Container:
      type: object
      required: [id, name, size, capacity, description, features, pricePerDay, imageUrl, order, isActive]
      properties:
        id: { type: integer }
        name: { type: string }
        size: { type: string }
        capacity: { type: string }
        description: { type: string }
        features:
          type: array
          items: { type: string }
        pricePerDay: { type: number }
        imageUrl: { type: string }
        order: { type: integer }
        isActive: { type: boolean }
    ContainerInput:
      type: object
      required: [name, size, capacity, description, features, pricePerDay, imageUrl, order]
      properties:
        name: { type: string }
        size: { type: string }
        capacity: { type: string }
        description: { type: string }
        features:
          type: array
          items: { type: string }
        pricePerDay: { type: number }
        imageUrl: { type: string }
        order: { type: integer }
        isActive: { type: boolean }
...
    Testimonial:
      type: object
      required: [id, clientName, company, content, rating, isActive]
      properties:
        id: { type: integer }
        clientName: { type: string }
        company: { type: string }
        content: { type: string }
        rating: { type: integer }
        avatarUrl: { type: ["string", "null"] }
        isActive: { type: boolean }
        createdAt: { type: string }
    TestimonialInput:
      type: object
      required: [clientName, company, content, rating]
      properties:
        clientName: { type: string }
        company: { type: string }
        content: { type: string }
        rating: { type: integer }
        avatarUrl: { type: ["string", "null"] }
        isActive: { type: boolean }
...
    ServiceRequest:
      type: object
      required: [id, clientName, phone, serviceType, containerSize, location, status, createdAt]
      properties:
        id: { type: integer }
        clientName: { type: string }
        phone: { type: string }
        email: { type: ["string", "null"] }
        serviceType: { type: string }
        containerSize: { type: string }
        location: { type: string }
        duration: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        status:
          type: string
          enum: [pending, in_progress, completed, cancelled]
        adminNotes: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    ServiceRequestInput:
      type: object
      required: [clientName, phone, serviceType, containerSize, location]
      properties:
        clientName: { type: string }
        phone: { type: string }
        email: { type: ["string", "null"] }
        serviceType: { type: string }
        containerSize: { type: string }
        location: { type: string }
        duration: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
    ServiceRequestUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, in_progress, completed, cancelled]
        adminNotes: { type: ["string", "null"] }
...