# lib/api-zod/src/generated/types/container.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Sabaik Almasa API
 * OpenAPI spec version: 0.1.0
 */

export interface Container {
  id: number;
  name: string;
  size: string;
  capacity: string;
  description: string;
  features: string[];
  pricePerDay: number;
  imageUrl: string;
  order: number;
  isActive: boolean;
}