# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Sabaik Almasa API
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: number;
  title: string;
  description: string;
  icon: string;
  /** @nullable */
  imageUrl?: string | null;
  order: number;
  isActive: boolean;
}