# lib/api-zod/src/generated/types/serviceRequest.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Sabaik Almasa API
 * OpenAPI spec version: 0.1.0
 */
import type { ServiceRequestStatus } from './serviceRequestStatus';

export interface ServiceRequest {
  id: number;
  clientName: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  serviceType: string;
  containerSize: string;
  location: string;
  /** @nullable */
  duration?: string | null;
  /** @nullable */
  notes?: string | null;
  status: ServiceRequestStatus;
  /** @nullable */
  adminNotes?: string | null;
  createdAt: string;
  updatedAt?: string;
}